requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.2  # Parquet support

# AWS SDK
boto3==1.34.21
botocore==1.31.85
awswrangler==3.5.2  # AWS Data Wrangler for S3/Glue

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
imbalanced-learn==0.11.0  # Handle class imbalance

# Experiment Tracking
mlflow==2.9.2

# API
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
python-multipart==0.0.6

# Data Validation
pandera==0.17.2  # DataFrame validation

# Monitoring & Drift Detection
evidently==0.4.12  # Drift monitoring

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
loguru==0.7.2  # Better logging

# Development
pytest==7.4.4
pytest-cov==4.1.0
black==23.12.1  # Code formatting
flake8==7.0.0  # Linting
ipykernel==6.28.0  # Jupyter support
jupyter==1.0.0
notebook==7.0.6

# Airflow (Optional - for orchestration)
# apache-airflow==2.8.0
# apache-airflow-providers-amazon==8.16.0

# Docker/Deployment
gunicorn==21.2.0
